moved rsn controller to its own and finalized the ulong userId and auth changes.

+++ RngHelpdesk.Contracts/Users/Queries/GetUserLifecycleHistoryResponse.cs
@@ -4,6 +4,6 @@ namespace RngHelpdesk.Contracts.Users.Queries;
 
 public sealed class GetUserLifecycleHistoryResponse
 {
-    public int UserId { get; init; }
+    public ulong UserId { get; init; }
     public IReadOnlyList<UserLifecycleHistoryItem> History { get; init; } = [];
 }
